Enhancement: - Changed the ClientService to fit to the new backend. Now we use only the http status codes to decide whether a request succeeded or not. - Changed everything that is related to ClientService calls. - Some bugfixes

+++ src/xcode/MRNBike/AddReportViewController.swift
@@ -197,7 +197,7 @@ class AddReportViewController: UIViewController, UITextFieldDelegate, UIGestureR
             type = "Dangerous"
         }
         
-        let timestamp = Int(NSDate().timeIntervalSince1970 * 1000)
+        let timestamp = Int64(NSDate().timeIntervalSince1970 * 1000)
         
         var annotations = mapView.annotations.filter { $0 !== mapView.userLocation }
         if annotations.count == 0 {
